graph -> canvas w/ optional grid lines
