+++ src/plugins/canvas/Plugin.tsx
