+++ src/plugins/canvas/languages/BasicJS/implementation.ts
@@ -1,0 +1,35 @@
+import { FromRuntimeMessage } from "../../messages";
+
+const createExports = (sendMessage: (message: FromRuntimeMessage) => void) => {
+  return Promise.resolve({
+    drawLine: (x1: number, y1: number, x2: number, y2: number) => {
+      sendMessage({ drawLine: { x1, y1, x2, y2 } });
+    },
+    drawPoint: (x: number, y: number) => {
+      sendMessage({ drawPoint: { x, y } });
+    },
+    drawCircle: (x: number, y: number, radius: number) => {
+      sendMessage({ drawCircle: { x, y, radius } });
+    },
+    setColor: (color: string) => {
+      sendMessage({ setColor: { color } });
+    },
+    drawVector: (x1: number, y1: number) => {
+      sendMessage({ drawVector: { x1, y1 } });
+    },
+    drawText: (text: string, x: number, y: number) => {
+      sendMessage({ drawText: { text, x, y } });
+    },
+    clear: () => {
+      sendMessage({ clear: true });
+    },
+    showGrid: (gridSize?: number, tickSize?: number) => {
+      sendMessage({ grid: { on: true, gridSize, tickSize } });
+    },
+    hideGrid: () => {
+      sendMessage({ grid: { on: false } });
+    },
+  });
+};
+
+export default createExports;
