+++ src/plugins/canvas/languages/BasicJS/starter-code/main.js
