+++ src/plugins/canvas/messages.ts
@@ -1,0 +1,37 @@
+export type FromRuntimeMessage = {
+  drawLine?: {
+    x1: number;
+    y1: number;
+    x2: number;
+    y2: number;
+  };
+  drawPoint?: {
+    x: number;
+    y: number;
+  };
+  drawCircle?: {
+    x: number;
+    y: number;
+    radius: number;
+  };
+  setColor?: {
+    color: string;
+  };
+  drawVector?: {
+    x1: number;
+    y1: number;
+  };
+  drawText?: {
+    text: string;
+    x: number;
+    y: number;
+  };
+  clear?: boolean;
+  grid?: {
+    on: boolean;
+    gridSize?: number;
+    tickSize?: number;
+  };
+};
+
+export type ToRuntimeMessage = {};
